+++ CHANGELOG.md
@@ -26,6 +26,7 @@ The project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html
   - Money Flow Index (MFI)
   - Midpoint Over Period (MOP)
   - Pivot Points (PivotPoints)
+  - Jurik Moving Average Average (JMA)
 - Fixes
   - Fixed Timeframe rounding to correctly handle days
   - Fixed OBV